feat: drag and drop and read data by document not collections
The core functionality of drag and drop has been implemented and we now are able to reorder cards within a section, but not move cards to other sections.

We now have a document inside the different collections with a field of tasks, in order to be able to update the array of tasks easily, i.e. being able to update a certain index of a task and then modifiy the tasks array and update the whole array itself.