feat: read tasks and handle dragndrop logic in Board component
In the parent component of the BoardColumn, Board itself, we know query for all the different type of tasks.

This allows us to read the data more efficiently, instead of having to do it in every Board Column.

The reordering logic within a list is still working as it should.