feat(useClickOutside): check if returned button ref was clicked
We now return a button, which can be used in order to tell the hook not to trigger the callback if that button was clicked, despite it being outside the containerRef.

We don't use a check function since it is not used anywhere in the project and rather have a boolean variable inside of the hook.

We return an object instead of an array with the refs from the hook because TypeScript understand those types somehow better.